requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==13.10
psycopg2-binary
feedparser
cachetools
speedtest-cli
cowsay
configparser